Create handler for key press functionality
